+++ src/server/services/ts/report.ts
@@ -2,33 +2,8 @@ import { runDbQuery } from "../db";
 import { generateReport as old_generateReport } from '../../holistic-approach/report-output';
 import { formData as mockedFormData } from '../../holistic-approach/hardcoded-mocks';
 import { type FormData } from '../../holistic-approach/io.types';
+import { FormDataType } from "@/types/report";
 
-// Johan's Proposal
-export type FormDataType = {
-  region?: "Europe" | "North America" | "Global"; // ...and more | default: Global
-  product?: "All products" | "Fine powder"; // ...and more | default: All products
-  valueChainStage?: {
-    mining?: boolean; // default: true
-    refining?: boolean; // default: true
-    firstUse?: boolean; // default: true
-    endUse?: boolean; // default: true
-    recycling?: boolean; // default: true
-  };
-  firstUseMode?:
-  | "ISIC sectorial analysis"
-  | "Representative Companies"
-  | "Average"; // default: ISIC sectorial analysis
-  contribution?: {
-    input?: boolean; // default: true
-    valueAdded?: boolean; // default: true
-  };
-  effect?: {
-    directEffect?: boolean; // default: true
-    firstRound?: boolean; // default: true
-    industrialSupport?: boolean; // default: true
-    incomeEffect?: boolean; // default: true
-  };
-};
 
 const LAST_YEAR = 2022; // TODO: This should be the last year available in the data from the supplier
